\name{easyrodbcaccess-package}
\alias{easyrodbcaccess-package}
\docType{package}
\title{Simple database interface for SQLite, MySQL and MS Access.}
\description{Simple database interface for MS Access. easyrodbcaccess is build on top of RODBC. It provides functions for reading, writing, listing and subsetting tables in MS Access. easyrodbcaccess allow to retrieve, write and subset database's tables in a similar way as data.frame's, using single square brackets db[,,,]. easyrodbcaccess internalise many database operations (otherwise tedious) such as rows and columns subsetting (no SQL needed) and connection and disconnection to / from databases.}
\details{
\tabular{ll}{Package: \tab easyrodbcaccess\cr
Version: \tab 0.7.0\cr
Date: \tab 2012-08-01\cr
Title: \tab Simple database interface for SQLite, MySQL and MS Access.\cr
Author: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Maintainer: \tab Julien MOEYS <Julien.Moeys@mark.slu.se>\cr
Depends: \tab R (>= 2.15.1), RODBC, easydb\cr
Suggests: \tab easyrodbcexcel, easyrsqlite, RODBC\cr
License: \tab AGPL-3\cr
URL: \tab http://easydb.r-forge.r-project.org/\cr}
}
\author{Julien MOEYS <Julien.Moeys@mark.slu.se>}

\keyword{ package }


